+++ discord/features/index.js
@@ -1,0 +1,9 @@
+const rawMessageHandler = require("./rawMessageHandler");
+const roles = require("./roles");
+const suggestions = require("./suggestions");
+
+module.exports = {
+	rawMessageHandler,
+	roles,
+	suggestions,
+};
