+++ discord/features/rawMessageHandler.js
@@ -1,0 +1,48 @@
+module.exports = client => {
+	client.on("raw", async packet => {
+		if (packet.t !== "MESSAGE_UPDATE") return;
+		// Grab the channel to check the message from
+		const channel = client.channels.get(packet.d.channel_id);
+		// There's no need to emit if the message is cached, because the event will fire anyway for that
+		if (channel.messages.has(packet.d.id)) return;
+		// Since we have confirmed the message is not cached, let's fetch it
+		const message = await channel.fetchMessage(packet.d.id);
+		// Emit the message
+		client.emit("messageUpdate", "", message);
+	});
+
+	client.on("raw", async packet => {
+		// We don't want this to run on unrelated packets
+		if (!["MESSAGE_REACTION_ADD", "MESSAGE_REACTION_REMOVE"].includes(packet.t))
+			return;
+		// Grab the channel to check the message from
+		const channel = client.channels.get(packet.d.channel_id);
+		// There's no need to emit if the message is cached, because the event will fire anyway for that
+		if (channel.messages.has(packet.d.message_id)) return;
+		// Since we have confirmed the message is not cached, let's fetch it
+		const message = await channel.fetchMessage(packet.d.message_id);
+		// Emojis can have identifiers of name:id format, so we have to account for that case as well
+		const emoji = packet.d.emoji.id
+			? `${packet.d.emoji.name}:${packet.d.emoji.id}`
+			: packet.d.emoji.name;
+		// This gives us the reaction we need to emit the event properly, in top of the message object
+		const reaction = message.reactions.get(emoji);
+		// Adds the currently reacting user to the reaction's users collection.
+		if (reaction)
+			reaction.users.set(packet.d.user_id, client.users.get(packet.d.user_id));
+		// Check which type of event it is before emitting
+		if (packet.t === "MESSAGE_REACTION_ADD") {
+			client.emit(
+				"messageReactionAdd",
+				reaction,
+				client.users.get(packet.d.user_id)
+			);
+		} else if (packet.t === "MESSAGE_REACTION_REMOVE") {
+			client.emit(
+				"messageReactionRemove",
+				reaction,
+				client.users.get(packet.d.user_id)
+			);
+		}
+	});
+};
