+++ helpers/isProduction.js
@@ -1,0 +1,7 @@
+const dotenv = require("dotenv");
+
+dotenv.config();
+
+const isProduction = process.env.ENV === "production";
+
+module.exports = isProduction;
